requirements.txt
anyio==4.7.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
courlan==1.3.2
cssselect==1.2.0
dateparser==1.2.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
html2text==2024.2.26
htmldate==1.9.2
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jusText==3.0.1
libmagic==1.0
lxml==5.3.0
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
readability==0.3.1
readability-lxml==0.8.1
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.5.0
tld==0.13
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.47.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
